rename auth, golint and govet
