+++ pkg/api/api_test.go
@@ -23,7 +23,9 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/sapcc/maia/pkg/keystone"
+	"encoding/json"
+	"github.com/databus23/goslo.policy"
+	"github.com/sapcc/maia/pkg/auth"
 	"github.com/sapcc/maia/pkg/storage"
 	"github.com/sapcc/maia/pkg/test"
 	"github.com/spf13/viper"
