+++ pkg/api/server.go
@@ -24,7 +24,7 @@ import (
 
 	"fmt"
 	"github.com/gorilla/mux"
-	"github.com/sapcc/maia/pkg/keystone"
+	"github.com/sapcc/maia/pkg/auth"
 	"github.com/sapcc/maia/pkg/storage"
 	"github.com/sapcc/maia/pkg/util"
 	"github.com/spf13/viper"
