+++ pkg/auth/interface.go
